Changed start argument for Str_PartAppend to size_t

+++ doomsday/engine/portable/include/m_string.h
@@ -63,7 +63,7 @@ void            Str_Set(ddstring_t* ds, const char* text);
 void            Str_Append(ddstring_t* ds, const char* appendText);
 void            Str_AppendChar(ddstring_t* ds, char ch);
 void            Str_Appendf(ddstring_t* ds, const char* format, ...) PRINTF_F(2,3);
-void            Str_PartAppend(ddstring_t* dest, const char* src, int start, size_t count);
+void            Str_PartAppend(ddstring_t* dest, const char* src, size_t start, size_t count);
 void            Str_Prepend(ddstring_t* ds, const char* prependText);
 size_t          Str_Length(const ddstring_t* ds);
 char*           Str_Text(const ddstring_t* ds);
